Move most event listener logic to actions/event-listeners
